+++ Sasquatch/SasquatchSwift/AppDelegate.swift
@@ -322,6 +322,20 @@ extension AppDelegate: DistributeDelegate {
     alert.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
     self.window?.rootViewController?.present(alert, animated: true)
   }
+
+  func distributeWillExitApp(_ distribute: Distribute) {
+    print("distributeWillExitApp callback invoked");
+    DispatchQueue.main.async {
+      let alertController = UIAlertController(title: NSLocalizedString("distribute_alert_willExit_title", tableName: "Sasquatch", comment: ""),
+              message: NSLocalizedString("distribute_alert_willExit_message", tableName: "Sasquatch", comment: ""),
+              preferredStyle: .alert)
+
+      // Show alert saying that the app is closing
+      self.window?.rootViewController?.present(alertController, animated: true)
+    }
+
+    sleep(7)
+  }
 }
 
 #endif