+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### App Center Distribute
 
 * **[Feature]** Add `distributeNoReleaseAvailable` callback to Distribute delegate.
+* **[Feature]** Add `distributeWillExitApp` callback to Distribute delegate.
 * **[Feature]** Add the Distribute module to Swift Package Manager. Please note that due to additional resources bundle, it requires Swift 5.3 (Xcode 12 and higher).
 
 ___